Move agents to 1 file
