+++ client/agents/CMakeLists.txt
@@ -2,8 +2,6 @@ add_library(
   agents
   STATIC
   agents.cpp
-  cma_core.cpp
-  cma_fec.cpp
 )
 
 target_include_directories(agents PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
